refactor: 폴더 구조 변경 및 경로 상수 추가
