+++ src/components/Footer/Footer.tsx
@@ -1,10 +1,9 @@
 import React from "react";
 import Image from "next/image";
-
 import Link from "next/link";
 import styled from "styled-components";
 import media from "@/styles/media";
-import SNSMenu from "../SNSMenu";
+import SNSMenu from "../common/SNSMenu";
 
 const Footer = () => {
   return (
