+++ src/components/common/Layout/Layout.tsx
@@ -1,5 +1,5 @@
-import Footer from "@/components/common/Footer";
-import Header from "@/components/common/Header";
+import Footer from "@/components/Footer";
+import Header from "@/components/Header";
 import React from "react";
 import styled from "styled-components";
