+++ src/constants/path.ts
@@ -1,0 +1,9 @@
+const PATHS = {
+  home: "/",
+  wine: {
+    search: "/wine/search",
+    detail: "/wine/[id]",
+  },
+};
+
+export default PATHS;
